Return a list of all possible ids

Return a list of all possible ids that can be used in the 'to' argument

cts_to(verbose = getOption("verbose"))

Arguments

verbose
logical; should a verbose output be printed on the console?

Value

a character vector.

Details

See also http://cts.fiehnlab.ucdavis.edu/services

References

Wohlgemuth, G., P. K. Haldiya, E. Willighagen, T. Kind, and O. Fiehn 2010The Chemical Translation Service -- a Web-Based Tool to Improve Standardization of Metabolomic Reports. Bioinformatics 26(20): 2647–2648.

See also

Examples

# \donttest{
cts_from()

#>   [1] "biocyc"                                                                                                                                     
#>   [2] "cas"                                                                                                                                        
#>   [3] "chebi"                                                                                                                                      
#>   [4] "chemical name"                                                                                                                              
#>   [5] "human metabolome database"                                                                                                                  
#>   [6] "inchikey"                                                                                                                                   
#>   [7] "kegg"                                                                                                                                       
#>   [8] "lmsd"                                                                                                                                       
#>   [9] "lipidmaps"                                                                                                                                  
#>  [10] "pubchem cid"                                                                                                                                
#>  [11] "pubchem sid"                                                                                                                                
#>  [12] "chemspider"                                                                                                                                 
#>  [13] "chemdb"                                                                                                                                     
#>  [14] "chembl"                                                                                                                                     
#>  [15] "chembank"                                                                                                                                   
#>  [16] "zinc"                                                                                                                                       
#>  [17] "broad institute"                                                                                                                            
#>  [18] "the scripps research institute molecular screening center"                                                                                  
#>  [19] "drugbank"                                                                                                                                   
#>  [20] "mmdb"                                                                                                                                       
#>  [21] "chemmol"                                                                                                                                    
#>  [22] "niaid"                                                                                                                                      
#>  [23] "southern research institute"                                                                                                                
#>  [24] "nih clinical collection"                                                                                                                    
#>  [25] "comparative toxicogenomics database"                                                                                                        
#>  [26] "molport"                                                                                                                                    
#>  [27] "smid"                                                                                                                                       
#>  [28] "chembridge"                                                                                                                                 
#>  [29] "glaxosmithkline (gsk)"                                                                                                                      
#>  [30] "ninds approved drug screening program"                                                                                                      
#>  [31] "nist"                                                                                                                                       
#>  [32] "scripdb"                                                                                                                                    
#>  [33] "southern research specialized biocontainment screening center"                                                                              
#>  [34] "timtec"                                                                                                                                     
#>  [35] "tox21"                                                                                                                                      
#>  [36] "web of science"                                                                                                                             
#>  [37] "abi chem"                                                                                                                                   
#>  [38] "center for chemical genomics, university of michigan"                                                                                       
#>  [39] "chemidplus"                                                                                                                                 
#>  [40] "sigma-aldrich"                                                                                                                              
#>  [41] "asinex"                                                                                                                                     
#>  [42] "chemblock"                                                                                                                                  
#>  [43] "discoverygate"                                                                                                                              
#>  [44] "specs"                                                                                                                                      
#>  [45] "ambinter"                                                                                                                                   
#>  [46] "emory university molecular libraries screening center"                                                                                      
#>  [47] "mlsmr"                                                                                                                                      
#>  [48] "vitas-m laboratory"                                                                                                                         
#>  [49] "aaa chemistry"                                                                                                                              
#>  [50] "chemfrog"                                                                                                                                   
#>  [51] "isoprenoids"                                                                                                                                
#>  [52] "ncgc"                                                                                                                                       
#>  [53] "acesobio"                                                                                                                                   
#>  [54] "aldrich"                                                                                                                                    
#>  [55] "ambit biosciences"                                                                                                                          
#>  [56] "clri (csir)"                                                                                                                                
#>  [57] "infarmatik"                                                                                                                                 
#>  [58] "is chemical technology"                                                                                                                     
#>  [59] "moli"                                                                                                                                       
#>  [60] "sigma"                                                                                                                                      
#>  [61] "tetrahedron scientific inc"                                                                                                                 
#>  [62] "thomson pharma"                                                                                                                             
#>  [63] "abblis chemicals"                                                                                                                           
#>  [64] "abbott labs"                                                                                                                                
#>  [65] "abmole bioscience"                                                                                                                          
#>  [66] "achemica"                                                                                                                                   
#>  [67] "acorn pharmatech"                                                                                                                           
#>  [68] "active biopharma"                                                                                                                           
#>  [69] "adooq bioscience"                                                                                                                           
#>  [70] "akos consulting & solutions"                                                                                                                
#>  [71] "ak scientific, inc. (aksci)"                                                                                                                
#>  [72] "alagar yadav, karpagam university"                                                                                                          
#>  [73] "alinda chemical"                                                                                                                            
#>  [74] "alsachim"                                                                                                                                   
#>  [75] "amadis chemical"                                                                                                                            
#>  [76] "amatye"                                                                                                                                     
#>  [77] "amicbase - antimicrobial activities"                                                                                                        
#>  [78] "angene chemical"                                                                                                                            
#>  [79] "angene international"                                                                                                                       
#>  [80] "ark pharm, inc."                                                                                                                            
#>  [81] "biochempartner"                                                                                                                             
#>  [82] "broadpharm"                                                                                                                                 
#>  [83] "chembase.cn"                                                                                                                                
#>  [84] "chembo"                                                                                                                                     
#>  [85] "chemicalize.org by chemaxon"                                                                                                                
#>  [86] "fragmenta"                                                                                                                                  
#>  [87] "anitha, department of bioinformatics, karpagam university"                                                                                  
#>  [88] "annker organics"                                                                                                                            
#>  [89] "anward"                                                                                                                                     
#>  [90] "apeiron synthesis"                                                                                                                          
#>  [91] "apexbio technology"                                                                                                                         
#>  [92] "apexmol"                                                                                                                                    
#>  [93] "aromsyn catalogue"                                                                                                                          
#>  [94] "aronis"                                                                                                                                     
#>  [95] "aurora fine chemicals llc"                                                                                                                  
#>  [96] "aurum pharmatech llc"                                                                                                                       
#>  [97] "avanti polar lipids"                                                                                                                        
#>  [98] "beijing advanced technology co, ltd"                                                                                                        
#>  [99] "bertin pharma"                                                                                                                              
#> [100] "bhaskar lab, department of zoology, sri venkateswara university"                                                                            
#> [101] "bidd"                                                                                                                                       
#> [102] "bind"                                                                                                                                       
#> [103] "bindingdb"                                                                                                                                  
#> [104] "biological magnetic resonance data bank (bmrb)"                                                                                             
#> [105] "bioprocess technology lab, department of microbiology, bharathidasan university"                                                            
#> [106] "biosynth"                                                                                                                                   
#> [107] "burnham center for chemical genomics"                                                                                                       
#> [108] "calbiochem"                                                                                                                                 
#> [109] "cambridge crystallographic data centre"                                                                                                     
#> [110] "capot"                                                                                                                                      
#> [111] "cayman chemical"                                                                                                                            
#> [112] "cc_pmlsc"                                                                                                                                   
#> [113] "chemexper chemical directory"                                                                                                               
#> [114] "chemical biology department, max planck institute of molecular physiology"                                                                  
#> [115] "chemscene"                                                                                                                                  
#> [116] "chemsynthesis"                                                                                                                              
#> [117] "chemtik"                                                                                                                                    
#> [118] "chiralblock biosciences"                                                                                                                    
#> [119] "circadian research, kay laboratory, university of california at san diego (ucsd)"                                                           
#> [120] "cmld-bu"                                                                                                                                    
#> [121] "columbia university molecular screening center"                                                                                             
#> [122] "creasyn finechem"                                                                                                                           
#> [123] "department of pharmacy, lmu"                                                                                                                
#> [124] "dtp/nci"                                                                                                                                    
#> [125] "edasa scientific compounds june 2013"                                                                                                       
#> [126] "emd biosciences"                                                                                                                            
#> [127] "enamine"                                                                                                                                    
#> [128] "ennopharm"                                                                                                                                  
#> [129] "epa dsstox"                                                                                                                                 
#> [130] "excenen pharmatech"                                                                                                                         
#> [131] "exchemistry"                                                                                                                                
#> [132] "finetech"                                                                                                                                   
#> [133] "finley and king labs, harvard medical school"                                                                                               
#> [134] "fluka"                                                                                                                                      
#> [135] "forechem"                                                                                                                                   
#> [136] "georganics"                                                                                                                                 
#> [137] "glida, gpcr-ligand database"                                                                                                                
#> [138] "gnf / scripps winzeler lab"                                                                                                                 
#> [139] "golm metabolome database (gmd), max planck institute of molecular plant physiology"                                                         
#> [140] "hangzhou apichem technology"                                                                                                                
#> [141] "hangzhou trylead chemical technology"                                                                                                       
#> [142] "hdh pharma"                                                                                                                                 
#> [143] "humgenex"                                                                                                                                   
#> [144] "ibch ras"                                                                                                                                   
#> [145] "ibm"                                                                                                                                        
#> [146] "iccb-longwood/nsrb screening facility, harvard medical school"                                                                              
#> [147] "immunology lab, department of biotechnology, calicut university"                                                                            
#> [148] "inhibitor 2"                                                                                                                                
#> [149] "insect molecular biology lab, department of environmental biotechnology, bharathidasan university"                                          
#> [150] "ithemba pharmaceuticals"                                                                                                                    
#> [151] "iuphar-db"                                                                                                                                  
#> [152] "jamson pharmachem technology"                                                                                                               
#> [153] "japan chemical substance dictionary (nikkaji)"                                                                                              
#> [154] "johns hopkins ion channel center"                                                                                                           
#> [155] "kingston chemistry"                                                                                                                         
#> [156] "kumgm"                                                                                                                                      
#> [157] "leadscope"                                                                                                                                  
#> [158] "medchemexpress mce"                                                                                                                         
#> [159] "micad"                                                                                                                                      
#> [160] "mic scientific"                                                                                                                             
#> [161] "milwaukee institute for drug discovery"                                                                                                     
#> [162] "molecular libraries program, specialized chemistry center, university of kansas"                                                            
#> [163] "mp biomedicals"                                                                                                                             
#> [164] "mtdp"                                                                                                                                       
#> [165] "nanjing pharmaceutical factory"                                                                                                             
#> [166] "nantong baihua bio-pharmaceutical co., ltd"                                                                                                 
#> [167] "nature chemical biology"                                                                                                                    
#> [168] "nature chemistry"                                                                                                                           
#> [169] "nature communications"                                                                                                                      
#> [170] "nist chemistry webbook"                                                                                                                     
#> [171] "nitric oxide research, national cancer institute (nci)"                                                                                     
#> [172] "nmmlsc"                                                                                                                                     
#> [173] "nmrshiftdb"                                                                                                                                 
#> [174] "novoseek"                                                                                                                                   
#> [175] "oakwood products"                                                                                                                           
#> [176] "orst small molecule screening center"                                                                                                       
#> [177] "p3 biosystems"                                                                                                                              
#> [178] "panache"                                                                                                                                    
#> [179] "paul baures"                                                                                                                                
#> [180] "pcmd"                                                                                                                                       
#> [181] "pdsp"                                                                                                                                       
#> [182] "penn-abs"                                                                                                                                   
#> [183] "pennchem-gam"                                                                                                                               
#> [184] "pfc"                                                                                                                                        
#> [185] "p.ravikumar, m.jeyam and g.shalini. biochematics division, bharathiar university"                                                           
#> [186] "priyadharshini sabarathinam angayarkanni murugesh palaniswamy"                                                                              
#> [187] "prous science drugs of the future"                                                                                                          
#> [188] "rangan lab, department of biotechnology, iit guwahati"                                                                                      
#> [189] "r&d chemicals"                                                                                                                              
#> [190] "r.sathishkumar, phytomatics laboratory, department of bioinformatics, bharathiar university"                                                
#> [191] "rschem, llc"                                                                                                                                
#> [192] "sastra university, quorum sensing and peptidomimetics laboratory"                                                                           
#> [193] "selleckbio"                                                                                                                                 
#> [194] "selleck chemicals"                                                                                                                          
#> [195] "sgcoxcompounds"                                                                                                                             
#> [196] "sgcstocompounds"                                                                                                                            
#> [197] "s.gurudeeban, t.ramanathan & k.satyavani, marine medicinal plant biotechnology laboratory, faculty of marine sciences, annamalai universtiy"
#> [198] "shanghai institute of organic chemistry"                                                                                                    
#> [199] "shanghai sinofluoro scientific company"                                                                                                     
#> [200] "sling consortium"                                                                                                                           
#> [201] "srmlsc"                                                                                                                                     
#> [202] "structural genomics consortium"                                                                                                             
#> [203] "surechem"                                                                                                                                   
#> [204] "synchem ohg"                                                                                                                                
#> [205] "syntechem"                                                                                                                                  
#> [206] "tci (tokyo chemical industry)"                                                                                                              
#> [207] "ten dijke lab, leiden university medical center"                                                                                            
#> [208] "therapeutic targets database"                                                                                                               
#> [209] "total toslab building-blocks"                                                                                                               
#> [210] "true pharmachem"                                                                                                                            
#> [211] "tyger scientific"                                                                                                                           
#> [212] "ucla molecular screening shared resource"                                                                                                   
#> [213] "um-bbd"                                                                                                                                     
#> [214] "unicarbkb"                                                                                                                                  
#> [215] "university of pittsburgh molecular library screening center"                                                                                
#> [216] "upcmld"                                                                                                                                     
#> [217] "vanderbilt screening center for gpcrs, ion channels and transporters"                                                                       
#> [218] "vanderbilt specialized chemistry center"                                                                                                    
#> [219] "vanderbilt university medical center"                                                                                                       
#> [220] "vit university"                                                                                                                             
#> [221] "watec laboratories"                                                                                                                         
#> [222] "watson international ltd"                                                                                                                   
#> [223] "xpharm"                                                                                                                                     
#> [224] "zancheng functional chemicals"                                                                                                              
#> [225] "zealing chemical"                                                                                                                           

# }