Retrieve Lite Entity (identifiers) from ChEBI

Returns a data.frame with a ChEBI entity ID (chebiid), a ChEBI entity name (chebiasciiname), a search score (searchscore) and stars (stars) using the SOAP protocol: https://www.ebi.ac.uk/chebi/webServices.do

get_chebiid(
  query,
  from = c("all", "chebi id", "chebi name", "definition", "name", "iupac name",
    "citations", "registry numbers", "manual xrefs", "automatic xrefs", "formula",
    "mass", "monoisotopic mass", "charge", "inchi", "inchikey", "smiles", "species"),
  match = c("all", "best", "first", "ask", "na"),
  max_res = 200,
  stars = c("all", "two only", "three only"),
  verbose = getOption("verbose"),
  ...
)

Arguments

query
character; search term.

from
character; type of input. "all" searches all types and "name" searches all names. Other options include 'chebi id', 'chebi name', 'definition', 'iupac name', 'citations', 'registry numbers', 'manual xrefs', 'automatic xrefs', 'formula', 'mass', 'monoisotopic mass','charge', 'inchi', 'inchikey', 'smiles', and 'species'

match
character; How should multiple hits be handled?, "all" all matches are returned, "best" the best matching (by the ChEBI searchscore) is returned, "ask" enters an interactive mode and the user is asked for input, "na" returns NA if multiple hits are found.

max_res
integer; maximum number of results to be retrieved from the web service

stars
character; "three only" restricts results to those manualy annotated by the ChEBI team.

verbose
logical; should a verbose output be printed on the console?

...
currently unused

Value

returns a list of data.frames containing a chebiid, a chebiasciiname, a searchscore and stars if matches were found. If not, data.frame(NA) is returned

References

Hastings J, Owen G, Dekker A, Ennis M, Kale N, Muthukrishnan V, Turner S, Swainston N, Mendes P, Steinbeck C. (2016). ChEBI in 2016: Improved services and an expanding collection of metabfolites. Nucleic Acids Res.

Hastings, J., de Matos, P., Dekker, A., Ennis, M., Harsha, B., Kale, N., Muthukrishnan, V., Owen, G., Turner, S., Williams, M., and Steinbeck, C. (2013) The ChEBI reference database and ontology for biologically relevant chemistry: enhancements for 2013. Nucleic Acids Res.

de Matos, P., Alcantara, R., Dekker, A., Ennis, M., Hastings, J., Haug, K., Spiteri, I., Turner, S., and Steinbeck, C. (2010) Chemical entities of biological interest: an update. Nucleic Acids Res. Degtyarenko, K., Hastings, J., de Matos, P., and Ennis, M. (2009). ChEBI: an open bioinformatics and cheminformatics resource. Current protocols in bioinformatics / editoral board, Andreas D. Baxevanis et al., Chapter 14.

Degtyarenko, K., de Matos, P., Ennis, M., Hastings, J., Zbinden, M., McNaught, A., Alcántara, R., Darsow, M., Guedj, M. and Ashburner, M. (2008) ChEBI: a database and ontology for chemical entities of biological interest. Nucleic Acids Res. 36, D344–D350.

Eduard Szöcs, Tamás Stirling, Eric R. Scott, Andreas Scharmüller, Ralf B. Schäfer (2020). webchem: An R Package to Retrieve Chemical Information from the Web. Journal of Statistical Software, 93(13). doi: 10.18637/jss.v093.i13 .

Examples

# \donttest{
# might fail if API is not available
get_chebiid('Glyphosate')

#> # A tibble: 7 x 5
#>   query      chebiid      chebiasciiname               searchscore entitystar
#>   <chr>      <chr>        <chr>                        <chr>       <chr>     
#> 1 Glyphosate CHEBI:27744  glyphosate                   0.94        3         
#> 2 Glyphosate CHEBI:133673 glyphosate(1-)               0.82        3         
#> 3 Glyphosate CHEBI:67052  glyphosate(2-)               0.66        3         
#> 4 Glyphosate CHEBI:81834  glyphosate-isopropylammonium 0.33        2         
#> 5 Glyphosate CHEBI:81835  Glyphosate-monoammonium      0.13        2         
#> 6 Glyphosate CHEBI:28812  (aminomethyl)phosphonic acid 0.06        3         
#> 7 Glyphosate CHEBI:142824 saflufenacil                 0.05        3         

get_chebiid('BPGDAMSIGCZZLK-UHFFFAOYSA-N')

#> # A tibble: 1 x 2
#>   query                       chebiid
#>   <chr>                       <chr>  
#> 1 BPGDAMSIGCZZLK-UHFFFAOYSA-N NA     

# multiple inputs
comp <- c('Iron', 'Aspirin', 'BPGDAMSIGCZZLK-UHFFFAOYSA-N')
get_chebiid(comp)

#> # A tibble: 219 x 5
#>    query chebiid      chebiasciiname                  searchscore entitystar
#>    <chr> <chr>        <chr>                           <chr>       <chr>     
#>  1 Iron  CHEBI:53438  iron(3+) sulfate                0.92        3         
#>  2 Iron  CHEBI:75899  iron(3+) sulfide                0.89        3         
#>  3 Iron  CHEBI:144421 iron(III) citrate               0.87        3         
#>  4 Iron  CHEBI:131372 iron(3+) phosphate tetrahydrate 0.86        3         
#>  5 Iron  CHEBI:131387 iron(3+) sulfate hydrate        0.86        3         
#>  6 Iron  CHEBI:75832  iron(2+) sulfate (anhydrous)    0.83        3         
#>  7 Iron  CHEBI:75896  iron sulfides                   0.76        3         
#>  8 Iron  CHEBI:18248  iron atom                       0.75        3         
#>  9 Iron  CHEBI:82664  iron(0)                         0.72        3         
#> 10 Iron  CHEBI:131371 iron(3+) phosphate              0.7         3         
#> # … with 209 more rows

# }