Retrieve retention indices from NIST

This function scrapes NIST for literature retention indices given CAS numbers as an input.

nist_ri(
  query,
  from = c("cas", "inchi", "inchikey", "name"),
  type = c("kovats", "linear", "alkane", "lee"),
  polarity = c("polar", "non-polar"),
  temp_prog = c("isothermal", "ramp", "custom"),
  cas = NULL,
  verbose = getOption("verbose")
)

Arguments

query
character; the search term

from
character; type of search term. can be one of "name", "inchi", "inchikey", or "cas". Using an identifier is preferred to "name" since NA is returned in the event of multiple matches to a query. Using an identifier other than a CAS number will cause this function to run slower as CAS numbers are used as internal identifiers by NIST.

type
Retention index type. One of "kovats", "linear", "alkane", or "lee". See details for more.

polarity
Column polarity. One of "polar" or "non-polar" to get RIs calculated for polar or non-polar columns.

temp_prog
Temperature program. One of "isothermal", "ramp", or "custom".

cas
deprecated. Use query instead.

verbose
logical; should a verbose output be printed on the console?

Value

returns a tibble of literature RIs with the following columns:

CAS is the CAS number

type is the column type, e.g. "capillary"

phase is the stationary phase (column phase)

RI is retention index

length is column length in meters

gas is the carrier gas used

substrate

diameter is the column diameter in mm

thickness is the phase thickness in µm

program. various columns depending on the value of temp_prog

reference is where this retention index was published

comment. I believe this denotes the database these data were aggregated from

Details

The types of retention indices included in NIST include Kovats ("kovats"), Van den Dool and Kratz ("linear"), normal alkane ("alkane"), and Lee ("lee"). Details about how these are calculated are available on the NIST website: https://webbook.nist.gov/chemistry/gc-ri/

Note

Copyright for NIST Standard Reference Data is governed by the Standard Reference Data Act, https://www.nist.gov/srd/public-law.

References

NIST Mass Spectrometry Data Center, William E. Wallace, director, "Retention Indices" in NIST Chemistry WebBook, NIST Standard Reference Database Number 69, Eds. P.J. Linstrom and W.G. Mallard, National Institute of Standards and Technology, Gaithersburg MD, 20899, doi: 10.18434/T4D303 .

See also

Examples

if (FALSE) {
myRIs <- nist_ri(c("78-70-6", "13474-59-4"), from = "cas", "linear",
"non-polar", "ramp")
}