Retrieve Complete Entity from ChEBI

Returns a list of Complete ChEBI entities. ChEBI data are parsed as data.frames ("properties", "chebiid_snd", "synonyms", "iupacnames", "formulae", "regnumbers", "citations", "dblinks", "parents", "children", "comments", "origins") or as a list ("chem_structure") in the list. The SOAP protocol is used https://www.ebi.ac.uk/chebi/webServices.do.

chebi_comp_entity(chebiid, verbose = getOption("verbose"), ...)

Arguments

chebiid
character; search term (i.e. chebiid).

verbose
logical; should a verbose output be printed on the console?

...
optional arguments

Value

returns a list of data.frames or lists containing a complete ChEBI entity

References

Hastings J, Owen G, Dekker A, Ennis M, Kale N, Muthukrishnan V, Turner S, Swainston N, Mendes P, Steinbeck C. (2016). ChEBI in 2016: Improved services and an expanding collection of metabolites. Nucleic Acids Res.

Hastings, J., de Matos, P., Dekker, A., Ennis, M., Harsha, B., Kale, N., Muthukrishnan, V., Owen, G., Turner, S., Williams, M., and Steinbeck, C. (2013) The ChEBI reference database and ontology for biologically relevant chemistry: enhancements for 2013. Nucleic Acids Res.

de Matos, P., Alcantara, R., Dekker, A., Ennis, M., Hastings, J., Haug, K., Spiteri, I., Turner, S., and Steinbeck, C. (2010) Chemical entities of biological interest: an update. Nucleic Acids Res. Degtyarenko, K., Hastings, J., de Matos, P., and Ennis, M. (2009). ChEBI: an open bioinformatics and cheminformatics resource. Current protocols in bioinformatics / editoral board, Andreas D. Baxevanis et al., Chapter 14.

Degtyarenko, K., de Matos, P., Ennis, M., Hastings, J., Zbinden, M., McNaught, A., Alcántara, R., Darsow, M., Guedj, M. and Ashburner, M. (2008) ChEBI: a database and ontology for chemical entities of biological interest. Nucleic Acids Res. 36, D344–D350.

Eduard Szöcs, Tamás Stirling, Eric R. Scott, Andreas Scharmüller, Ralf B. Schäfer (2020). webchem: An R Package to Retrieve Chemical Information from the Web. Journal of Statistical Software, 93(13). doi: 10.18637/jss.v093.i13 .

Examples

# \donttest{
# might fail if API is not available
chebi_comp_entity('CHEBI:27744')

#> $`CHEBI:27744`
#> $`CHEBI:27744`$properties
#>       chebiid chebiasciiname
#> 1 CHEBI:27744     glyphosate
#>                                                                                                                                                                                                                                                                                          definition
#> 1 A phosphonic acid resulting from the formal oxidative coupling of the methyl group of methylphosphonic acid with the amino group of glycine. It is one of the most commonly used herbicdes worldwide, and the only one to target the enzyme 5-enolpyruvyl-3-shikimate phosphate synthase (EPSPS).
#>    status             smiles
#> 1 CHECKED OC(=O)CNCP(O)(O)=O
#>                                                                 inchi
#> 1 InChI=1S/C3H8NO5P/c5-3(6)1-4-2-10(7,8)9/h4H,1-2H2,(H,5,6)(H2,7,8,9)
#>                      inchikey charge      mass monoisotopicmass entitystar
#> 1 XDDAORKBJWWYJS-UHFFFAOYSA-N      0 169.07310        169.01401          3
#> 
#> $`CHEBI:27744`$chebiid_snd
#>      chebiids
#> 1  CHEBI:5510
#> 2 CHEBI:24423
#> 3 CHEBI:43013
#> 
#> $`CHEBI:27744`$chem_structure
#> $`CHEBI:27744`$chem_structure[[1]]
#> $`CHEBI:27744`$chem_structure[[1]]$structure
#> $`CHEBI:27744`$chem_structure[[1]]$structure[[1]]
#> [1] "\n  Marvin  01291313162D          \n\n 10  9  0  0  0  0            999 V2000\n   15.6197  -10.4907    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n   15.6197  -11.8188    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   16.7788  -12.4874    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n   14.4609  -12.4874    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   13.3112  -11.8188    0.0000 N   0  0  0  0  0  0  0  0  0  0  0  0\n   12.1614  -12.4874    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   11.0026  -11.8188    0.0000 P   0  0  0  0  0  0  0  0  0  0  0  0\n   11.0086  -10.4958    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n    9.8548  -12.4988    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n   10.6560  -13.1077    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n  1  2  2  0  0  0  0\n  2  3  1  0  0  0  0\n  2  4  1  0  0  0  0\n  4  5  1  0  0  0  0\n  5  6  1  0  0  0  0\n  6  7  1  0  0  0  0\n  7  8  2  0  0  0  0\n  7  9  1  0  0  0  0\n  7 10  1  0  0  0  0\nM  END\n"
#> 
#> 
#> $`CHEBI:27744`$chem_structure[[1]]$type
#> $`CHEBI:27744`$chem_structure[[1]]$type[[1]]
#> [1] "mol"
#> 
#> 
#> $`CHEBI:27744`$chem_structure[[1]]$dimension
#> $`CHEBI:27744`$chem_structure[[1]]$dimension[[1]]
#> [1] "2D"
#> 
#> 
#> $`CHEBI:27744`$chem_structure[[1]]$defaultStructure
#> $`CHEBI:27744`$chem_structure[[1]]$defaultStructure[[1]]
#> [1] "true"
#> 
#> 
#> 
#> $`CHEBI:27744`$chem_structure[[2]]
#> $`CHEBI:27744`$chem_structure[[2]]$structure
#> $`CHEBI:27744`$chem_structure[[2]]$structure[[1]]
#> [1] "ChEBI\n  Marvin  01291313172D          \n\n 10  9  0  0  0  0            999 V2000\n    9.1576   -5.8060    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n    8.7451   -6.5204    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n    9.1576   -7.2349    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n    7.9201   -6.5204    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n    7.5076   -7.2349    0.0000 N   0  0  0  0  0  0  0  0  0  0  0  0\n    6.6826   -7.2349    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n    6.2701   -7.9493    0.0000 P   0  0  0  0  0  0  0  0  0  0  0  0\n    6.9846   -8.3618    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n    5.8576   -8.6639    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n    5.5556   -7.5368    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n  1  2  2  0  0  0  0\n  2  3  1  0  0  0  0\n  2  4  1  0  0  0  0\n  4  5  1  0  0  0  0\n  5  6  1  0  0  0  0\n  6  7  1  0  0  0  0\n  7  8  1  0  0  0  0\n  7  9  1  0  0  0  0\n  7 10  2  0  0  0  0\nM  END\n"
#> 
#> 
#> $`CHEBI:27744`$chem_structure[[2]]$type
#> $`CHEBI:27744`$chem_structure[[2]]$type[[1]]
#> [1] "mol"
#> 
#> 
#> $`CHEBI:27744`$chem_structure[[2]]$dimension
#> $`CHEBI:27744`$chem_structure[[2]]$dimension[[1]]
#> [1] "2D"
#> 
#> 
#> $`CHEBI:27744`$chem_structure[[2]]$defaultStructure
#> $`CHEBI:27744`$chem_structure[[2]]$defaultStructure[[1]]
#> [1] "false"
#> 
#> 
#> 
#> 
#> $`CHEBI:27744`$synonyms
#>         data       type        source
#> 1 Glyphosate    SYNONYM KEGG COMPOUND
#> 2    Roundup BRAND NAME KEGG COMPOUND
#> 
#> $`CHEBI:27744`$iupacnames
#>                         data       type source
#> 1 N-(phosphonomethyl)glycine IUPAC NAME  IUPAC
#> 
#> $`CHEBI:27744`$formulae
#>       data        source
#> 1 C3H8NO5P         ChEBI
#> 2 C3H8NO5P KEGG COMPOUND
#> 
#> $`CHEBI:27744`$regnumbers
#>        data                      type                 source
#> 1 1071-83-6       CAS Registry Number          KEGG COMPOUND
#> 2 1071-83-6       CAS Registry Number Alan Wood's Pesticides
#> 3 1071-83-6       CAS Registry Number             ChemIDplus
#> 4   2045054 Beilstein Registry Number              Beilstein
#> 5    279222    Gmelin Registry Number                 Gmelin
#> 
#> $`CHEBI:27744`$citations
#>       data            type     source
#> 1 27758090 PubMed citation Europe PMC
#> 2 28266132 PubMed citation Europe PMC
#> 3 28474816 PubMed citation Europe PMC
#> 4 28643882 PubMed citation Europe PMC
#> 5 28711546 PubMed citation Europe PMC
#> 6 30471482 PubMed citation Europe PMC
#> 7 30875550 PubMed citation Europe PMC
#> 8 31030151 PubMed citation Europe PMC
#> 
#> $`CHEBI:27744`$dblinks
#>         data                    type
#> 1        373          PPDB accession
#> 2      c0134           UM-BBD compID
#> 3     C01705 KEGG COMPOUND accession
#> 4    DB04539      DrugBank accession
#> 5 glyphosate    Pesticides accession
#> 6 Glyphosate     Wikipedia accession
#> 7        GPF      PDBeChem accession
#> 8        GPJ      PDBeChem accession
#> 
#> $`CHEBI:27744`$parents
#>                                                              chebiName
#> 1                                                       glyphosate(1-)
#> 2                                                         agrochemical
#> 3                                                      phosphonic acid
#> 4                                                   glycine derivative
#> 5                                                       glyphosate(2-)
#> 6 EC 2.5.1.19 (3-phosphoshikimate 1-carboxyvinyltransferase) inhibitor
#> 7                                                            herbicide
#>        chebiId                 type  status cyclicRelationship
#> 1 CHEBI:133673 is conjugate acid of CHECKED               true
#> 2  CHEBI:33286             has role CHECKED              false
#> 3  CHEBI:44976                 is a CHECKED              false
#> 4  CHEBI:24373                 is a CHECKED              false
#> 5  CHEBI:67052 is conjugate acid of CHECKED               true
#> 6  CHEBI:20569             has role      OK              false
#> 7  CHEBI:24527             has role CHECKED              false
#> 
#> $`CHEBI:27744`$children
#>        chebiName      chebiId                 type  status cyclicRelationship
#> 1 glyphosate(1-) CHEBI:133673 is conjugate base of CHECKED               true
#> 2 glyphosate(2-)  CHEBI:67052 is conjugate base of CHECKED               true
#> 
#> $`CHEBI:27744`$comments
#> data frame with 0 columns and 0 rows
#> 
#> $`CHEBI:27744`$origins
#> data frame with 0 columns and 0 rows
#> 
#> 
#> attr(,"class")
#> [1] "chebi_comp_entity" "list"             

# multiple inputs
comp <- c('CHEBI:27744', 'CHEBI:27744')
chebi_comp_entity(comp)

#> $`CHEBI:27744`
#> $`CHEBI:27744`$properties
#>       chebiid chebiasciiname
#> 1 CHEBI:27744     glyphosate
#>                                                                                                                                                                                                                                                                                          definition
#> 1 A phosphonic acid resulting from the formal oxidative coupling of the methyl group of methylphosphonic acid with the amino group of glycine. It is one of the most commonly used herbicdes worldwide, and the only one to target the enzyme 5-enolpyruvyl-3-shikimate phosphate synthase (EPSPS).
#>    status             smiles
#> 1 CHECKED OC(=O)CNCP(O)(O)=O
#>                                                                 inchi
#> 1 InChI=1S/C3H8NO5P/c5-3(6)1-4-2-10(7,8)9/h4H,1-2H2,(H,5,6)(H2,7,8,9)
#>                      inchikey charge      mass monoisotopicmass entitystar
#> 1 XDDAORKBJWWYJS-UHFFFAOYSA-N      0 169.07310        169.01401          3
#> 
#> $`CHEBI:27744`$chebiid_snd
#>      chebiids
#> 1  CHEBI:5510
#> 2 CHEBI:24423
#> 3 CHEBI:43013
#> 
#> $`CHEBI:27744`$chem_structure
#> $`CHEBI:27744`$chem_structure[[1]]
#> $`CHEBI:27744`$chem_structure[[1]]$structure
#> $`CHEBI:27744`$chem_structure[[1]]$structure[[1]]
#> [1] "\n  Marvin  01291313162D          \n\n 10  9  0  0  0  0            999 V2000\n   15.6197  -10.4907    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n   15.6197  -11.8188    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   16.7788  -12.4874    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n   14.4609  -12.4874    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   13.3112  -11.8188    0.0000 N   0  0  0  0  0  0  0  0  0  0  0  0\n   12.1614  -12.4874    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   11.0026  -11.8188    0.0000 P   0  0  0  0  0  0  0  0  0  0  0  0\n   11.0086  -10.4958    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n    9.8548  -12.4988    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n   10.6560  -13.1077    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n  1  2  2  0  0  0  0\n  2  3  1  0  0  0  0\n  2  4  1  0  0  0  0\n  4  5  1  0  0  0  0\n  5  6  1  0  0  0  0\n  6  7  1  0  0  0  0\n  7  8  2  0  0  0  0\n  7  9  1  0  0  0  0\n  7 10  1  0  0  0  0\nM  END\n"
#> 
#> 
#> $`CHEBI:27744`$chem_structure[[1]]$type
#> $`CHEBI:27744`$chem_structure[[1]]$type[[1]]
#> [1] "mol"
#> 
#> 
#> $`CHEBI:27744`$chem_structure[[1]]$dimension
#> $`CHEBI:27744`$chem_structure[[1]]$dimension[[1]]
#> [1] "2D"
#> 
#> 
#> $`CHEBI:27744`$chem_structure[[1]]$defaultStructure
#> $`CHEBI:27744`$chem_structure[[1]]$defaultStructure[[1]]
#> [1] "true"
#> 
#> 
#> 
#> $`CHEBI:27744`$chem_structure[[2]]
#> $`CHEBI:27744`$chem_structure[[2]]$structure
#> $`CHEBI:27744`$chem_structure[[2]]$structure[[1]]
#> [1] "ChEBI\n  Marvin  01291313172D          \n\n 10  9  0  0  0  0            999 V2000\n    9.1576   -5.8060    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n    8.7451   -6.5204    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n    9.1576   -7.2349    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n    7.9201   -6.5204    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n    7.5076   -7.2349    0.0000 N   0  0  0  0  0  0  0  0  0  0  0  0\n    6.6826   -7.2349    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n    6.2701   -7.9493    0.0000 P   0  0  0  0  0  0  0  0  0  0  0  0\n    6.9846   -8.3618    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n    5.8576   -8.6639    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n    5.5556   -7.5368    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n  1  2  2  0  0  0  0\n  2  3  1  0  0  0  0\n  2  4  1  0  0  0  0\n  4  5  1  0  0  0  0\n  5  6  1  0  0  0  0\n  6  7  1  0  0  0  0\n  7  8  1  0  0  0  0\n  7  9  1  0  0  0  0\n  7 10  2  0  0  0  0\nM  END\n"
#> 
#> 
#> $`CHEBI:27744`$chem_structure[[2]]$type
#> $`CHEBI:27744`$chem_structure[[2]]$type[[1]]
#> [1] "mol"
#> 
#> 
#> $`CHEBI:27744`$chem_structure[[2]]$dimension
#> $`CHEBI:27744`$chem_structure[[2]]$dimension[[1]]
#> [1] "2D"
#> 
#> 
#> $`CHEBI:27744`$chem_structure[[2]]$defaultStructure
#> $`CHEBI:27744`$chem_structure[[2]]$defaultStructure[[1]]
#> [1] "false"
#> 
#> 
#> 
#> 
#> $`CHEBI:27744`$synonyms
#>         data       type        source
#> 1 Glyphosate    SYNONYM KEGG COMPOUND
#> 2    Roundup BRAND NAME KEGG COMPOUND
#> 
#> $`CHEBI:27744`$iupacnames
#>                         data       type source
#> 1 N-(phosphonomethyl)glycine IUPAC NAME  IUPAC
#> 
#> $`CHEBI:27744`$formulae
#>       data        source
#> 1 C3H8NO5P         ChEBI
#> 2 C3H8NO5P KEGG COMPOUND
#> 
#> $`CHEBI:27744`$regnumbers
#>        data                      type                 source
#> 1 1071-83-6       CAS Registry Number          KEGG COMPOUND
#> 2 1071-83-6       CAS Registry Number Alan Wood's Pesticides
#> 3 1071-83-6       CAS Registry Number             ChemIDplus
#> 4   2045054 Beilstein Registry Number              Beilstein
#> 5    279222    Gmelin Registry Number                 Gmelin
#> 
#> $`CHEBI:27744`$citations
#>       data            type     source
#> 1 27758090 PubMed citation Europe PMC
#> 2 28266132 PubMed citation Europe PMC
#> 3 28474816 PubMed citation Europe PMC
#> 4 28643882 PubMed citation Europe PMC
#> 5 28711546 PubMed citation Europe PMC
#> 6 30471482 PubMed citation Europe PMC
#> 7 30875550 PubMed citation Europe PMC
#> 8 31030151 PubMed citation Europe PMC
#> 
#> $`CHEBI:27744`$dblinks
#>         data                    type
#> 1        373          PPDB accession
#> 2      c0134           UM-BBD compID
#> 3     C01705 KEGG COMPOUND accession
#> 4    DB04539      DrugBank accession
#> 5 glyphosate    Pesticides accession
#> 6 Glyphosate     Wikipedia accession
#> 7        GPF      PDBeChem accession
#> 8        GPJ      PDBeChem accession
#> 
#> $`CHEBI:27744`$parents
#>                                                              chebiName
#> 1                                                       glyphosate(1-)
#> 2                                                         agrochemical
#> 3                                                      phosphonic acid
#> 4                                                   glycine derivative
#> 5                                                       glyphosate(2-)
#> 6 EC 2.5.1.19 (3-phosphoshikimate 1-carboxyvinyltransferase) inhibitor
#> 7                                                            herbicide
#>        chebiId                 type  status cyclicRelationship
#> 1 CHEBI:133673 is conjugate acid of CHECKED               true
#> 2  CHEBI:33286             has role CHECKED              false
#> 3  CHEBI:44976                 is a CHECKED              false
#> 4  CHEBI:24373                 is a CHECKED              false
#> 5  CHEBI:67052 is conjugate acid of CHECKED               true
#> 6  CHEBI:20569             has role      OK              false
#> 7  CHEBI:24527             has role CHECKED              false
#> 
#> $`CHEBI:27744`$children
#>        chebiName      chebiId                 type  status cyclicRelationship
#> 1 glyphosate(1-) CHEBI:133673 is conjugate base of CHECKED               true
#> 2 glyphosate(2-)  CHEBI:67052 is conjugate base of CHECKED               true
#> 
#> $`CHEBI:27744`$comments
#> data frame with 0 columns and 0 rows
#> 
#> $`CHEBI:27744`$origins
#> data frame with 0 columns and 0 rows
#> 
#> 
#> $`CHEBI:27744`
#> $`CHEBI:27744`$properties
#>       chebiid chebiasciiname
#> 1 CHEBI:27744     glyphosate
#>                                                                                                                                                                                                                                                                                          definition
#> 1 A phosphonic acid resulting from the formal oxidative coupling of the methyl group of methylphosphonic acid with the amino group of glycine. It is one of the most commonly used herbicdes worldwide, and the only one to target the enzyme 5-enolpyruvyl-3-shikimate phosphate synthase (EPSPS).
#>    status             smiles
#> 1 CHECKED OC(=O)CNCP(O)(O)=O
#>                                                                 inchi
#> 1 InChI=1S/C3H8NO5P/c5-3(6)1-4-2-10(7,8)9/h4H,1-2H2,(H,5,6)(H2,7,8,9)
#>                      inchikey charge      mass monoisotopicmass entitystar
#> 1 XDDAORKBJWWYJS-UHFFFAOYSA-N      0 169.07310        169.01401          3
#> 
#> $`CHEBI:27744`$chebiid_snd
#>      chebiids
#> 1  CHEBI:5510
#> 2 CHEBI:24423
#> 3 CHEBI:43013
#> 
#> $`CHEBI:27744`$chem_structure
#> $`CHEBI:27744`$chem_structure[[1]]
#> $`CHEBI:27744`$chem_structure[[1]]$structure
#> $`CHEBI:27744`$chem_structure[[1]]$structure[[1]]
#> [1] "\n  Marvin  01291313162D          \n\n 10  9  0  0  0  0            999 V2000\n   15.6197  -10.4907    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n   15.6197  -11.8188    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   16.7788  -12.4874    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n   14.4609  -12.4874    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   13.3112  -11.8188    0.0000 N   0  0  0  0  0  0  0  0  0  0  0  0\n   12.1614  -12.4874    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   11.0026  -11.8188    0.0000 P   0  0  0  0  0  0  0  0  0  0  0  0\n   11.0086  -10.4958    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n    9.8548  -12.4988    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n   10.6560  -13.1077    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n  1  2  2  0  0  0  0\n  2  3  1  0  0  0  0\n  2  4  1  0  0  0  0\n  4  5  1  0  0  0  0\n  5  6  1  0  0  0  0\n  6  7  1  0  0  0  0\n  7  8  2  0  0  0  0\n  7  9  1  0  0  0  0\n  7 10  1  0  0  0  0\nM  END\n"
#> 
#> 
#> $`CHEBI:27744`$chem_structure[[1]]$type
#> $`CHEBI:27744`$chem_structure[[1]]$type[[1]]
#> [1] "mol"
#> 
#> 
#> $`CHEBI:27744`$chem_structure[[1]]$dimension
#> $`CHEBI:27744`$chem_structure[[1]]$dimension[[1]]
#> [1] "2D"
#> 
#> 
#> $`CHEBI:27744`$chem_structure[[1]]$defaultStructure
#> $`CHEBI:27744`$chem_structure[[1]]$defaultStructure[[1]]
#> [1] "true"
#> 
#> 
#> 
#> $`CHEBI:27744`$chem_structure[[2]]
#> $`CHEBI:27744`$chem_structure[[2]]$structure
#> $`CHEBI:27744`$chem_structure[[2]]$structure[[1]]
#> [1] "ChEBI\n  Marvin  01291313172D          \n\n 10  9  0  0  0  0            999 V2000\n    9.1576   -5.8060    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n    8.7451   -6.5204    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n    9.1576   -7.2349    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n    7.9201   -6.5204    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n    7.5076   -7.2349    0.0000 N   0  0  0  0  0  0  0  0  0  0  0  0\n    6.6826   -7.2349    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n    6.2701   -7.9493    0.0000 P   0  0  0  0  0  0  0  0  0  0  0  0\n    6.9846   -8.3618    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n    5.8576   -8.6639    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n    5.5556   -7.5368    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n  1  2  2  0  0  0  0\n  2  3  1  0  0  0  0\n  2  4  1  0  0  0  0\n  4  5  1  0  0  0  0\n  5  6  1  0  0  0  0\n  6  7  1  0  0  0  0\n  7  8  1  0  0  0  0\n  7  9  1  0  0  0  0\n  7 10  2  0  0  0  0\nM  END\n"
#> 
#> 
#> $`CHEBI:27744`$chem_structure[[2]]$type
#> $`CHEBI:27744`$chem_structure[[2]]$type[[1]]
#> [1] "mol"
#> 
#> 
#> $`CHEBI:27744`$chem_structure[[2]]$dimension
#> $`CHEBI:27744`$chem_structure[[2]]$dimension[[1]]
#> [1] "2D"
#> 
#> 
#> $`CHEBI:27744`$chem_structure[[2]]$defaultStructure
#> $`CHEBI:27744`$chem_structure[[2]]$defaultStructure[[1]]
#> [1] "false"
#> 
#> 
#> 
#> 
#> $`CHEBI:27744`$synonyms
#>         data       type        source
#> 1 Glyphosate    SYNONYM KEGG COMPOUND
#> 2    Roundup BRAND NAME KEGG COMPOUND
#> 
#> $`CHEBI:27744`$iupacnames
#>                         data       type source
#> 1 N-(phosphonomethyl)glycine IUPAC NAME  IUPAC
#> 
#> $`CHEBI:27744`$formulae
#>       data        source
#> 1 C3H8NO5P         ChEBI
#> 2 C3H8NO5P KEGG COMPOUND
#> 
#> $`CHEBI:27744`$regnumbers
#>        data                      type                 source
#> 1 1071-83-6       CAS Registry Number          KEGG COMPOUND
#> 2 1071-83-6       CAS Registry Number Alan Wood's Pesticides
#> 3 1071-83-6       CAS Registry Number             ChemIDplus
#> 4   2045054 Beilstein Registry Number              Beilstein
#> 5    279222    Gmelin Registry Number                 Gmelin
#> 
#> $`CHEBI:27744`$citations
#>       data            type     source
#> 1 27758090 PubMed citation Europe PMC
#> 2 28266132 PubMed citation Europe PMC
#> 3 28474816 PubMed citation Europe PMC
#> 4 28643882 PubMed citation Europe PMC
#> 5 28711546 PubMed citation Europe PMC
#> 6 30471482 PubMed citation Europe PMC
#> 7 30875550 PubMed citation Europe PMC
#> 8 31030151 PubMed citation Europe PMC
#> 
#> $`CHEBI:27744`$dblinks
#>         data                    type
#> 1        373          PPDB accession
#> 2      c0134           UM-BBD compID
#> 3     C01705 KEGG COMPOUND accession
#> 4    DB04539      DrugBank accession
#> 5 glyphosate    Pesticides accession
#> 6 Glyphosate     Wikipedia accession
#> 7        GPF      PDBeChem accession
#> 8        GPJ      PDBeChem accession
#> 
#> $`CHEBI:27744`$parents
#>                                                              chebiName
#> 1                                                       glyphosate(1-)
#> 2                                                         agrochemical
#> 3                                                      phosphonic acid
#> 4                                                   glycine derivative
#> 5                                                       glyphosate(2-)
#> 6 EC 2.5.1.19 (3-phosphoshikimate 1-carboxyvinyltransferase) inhibitor
#> 7                                                            herbicide
#>        chebiId                 type  status cyclicRelationship
#> 1 CHEBI:133673 is conjugate acid of CHECKED               true
#> 2  CHEBI:33286             has role CHECKED              false
#> 3  CHEBI:44976                 is a CHECKED              false
#> 4  CHEBI:24373                 is a CHECKED              false
#> 5  CHEBI:67052 is conjugate acid of CHECKED               true
#> 6  CHEBI:20569             has role      OK              false
#> 7  CHEBI:24527             has role CHECKED              false
#> 
#> $`CHEBI:27744`$children
#>        chebiName      chebiId                 type  status cyclicRelationship
#> 1 glyphosate(1-) CHEBI:133673 is conjugate base of CHECKED               true
#> 2 glyphosate(2-)  CHEBI:67052 is conjugate base of CHECKED               true
#> 
#> $`CHEBI:27744`$comments
#> data frame with 0 columns and 0 rows
#> 
#> $`CHEBI:27744`$origins
#> data frame with 0 columns and 0 rows
#> 
#> 
#> attr(,"class")
#> [1] "chebi_comp_entity" "list"             

# }