Retrieve Pubchem Compound ID (CID)

Retrieve compound IDs (CIDs) from PubChem.

get_cid(
  query,
  from = "name",
  domain = c("compound", "substance", "assay"),
  match = c("all", "first", "ask", "na"),
  verbose = getOption("verbose"),
  arg = NULL,
  first = NULL,
  ...
)

Arguments

query
character; search term, one or more compounds.

from
character; type of input. See details for more information.

domain
character; query domain, can be one of "compound", "substance", "assay".

match
character; How should multiple hits be handled?, "all" all matches are returned, "first" the first matching is returned, "ask" enters an interactive mode and the user is asked for input, "na" returns NA if multiple hits are found.

verbose
logical; should a verbose output be printed on the console?

arg
character; optinal arguments like "name_type=word" to match individual words.

first
deprecated. Use `match` instead.

...
currently unused.

Value

a tibble.

Details

Valid values for the from argument depend on the domain:

compound: "name", "smiles", "inchi", "inchikey", "formula", "sdf", <xref>, <structure search>, <fast search>.

substance: "name", "sid", <xref>, "sourceid/<source id>" or "sourceall".

assay: "aid", <assay target>.

<structure search> is assembled as "substructure | superstructure | similarity | identity / smiles | inchi | sdf | cid", e.g. from = "substructure/smiles".

<xref> is assembled as "xref/{RegistryID | RN | PubMedID | MMDBID | ProteinGI, NucleotideGI | TaxonomyID | MIMID | GeneID | ProbeID | PatentID}", e.g. from = "xref/RN" will query by CAS RN.

<fast search> is either fastformula or it is assembled as "fastidentity | fastsimilarity_2d | fastsimilarity_3d | fastsubstructure | fastsuperstructure/smiles | smarts | inchi | sdf | cid", e.g. from = "fastidentity/smiles".

<source id> is any valid PubChem Data Source ID. When from = "sourceid/<source id>", the query is the ID of the substance in the depositor's database.

If from = "sourceall" the query is one or more valid Pubchem depositor names. Depositor names are not case sensitive.

Depositor names and Data Source IDs can be found at https://pubchem.ncbi.nlm.nih.gov/sources/.

<assay target> is assembled as "target/{gi | proteinname | geneid | genesymbol | accession}", e.g. from = "target/geneid" will query by GeneID.

Note

Please respect the Terms and Conditions of the National Library of Medicine, https://www.nlm.nih.gov/databases/download.html the data usage policies of National Center for Biotechnology Information, https://www.ncbi.nlm.nih.gov/home/about/policies/, https://pubchemdocs.ncbi.nlm.nih.gov/programmatic-access, and the data usage policies of the indicidual data sources https://pubchem.ncbi.nlm.nih.gov/sources/.

References

Wang, Y., J. Xiao, T. O. Suzek, et al. 2009 PubChem: A Public Information System for Analyzing Bioactivities of Small Molecules. Nucleic Acids Research 37: 623–633.

Kim, Sunghwan, Paul A. Thiessen, Evan E. Bolton, et al. 2016 PubChem Substance and Compound Databases. Nucleic Acids Research 44(D1): D1202–D1213.

Kim, S., Thiessen, P. A., Bolton, E. E., & Bryant, S. H. (2015). PUG-SOAP and PUG-REST: web services for programmatic access to chemical information in PubChem. Nucleic acids research, gkv396.

Eduard Szöcs, Tamás Stirling, Eric R. Scott, Andreas Scharmüller, Ralf B. Schäfer (2020). webchem: An R Package to Retrieve Chemical Information from the Web. Journal of Statistical Software, 93(13). doi: 10.18637/jss.v093.i13 .

Examples

# \donttest{
# might fail if API is not available
get_cid("Triclosan")

#> # A tibble: 1 x 2
#>   query     cid  
#>   <chr>     <chr>
#> 1 Triclosan 5564 

get_cid("Triclosan", arg = "name_type=word")

#> # A tibble: 40 x 2
#>    query     cid      
#>    <chr>     <chr>    
#>  1 Triclosan 5564     
#>  2 Triclosan 627458   
#>  3 Triclosan 45040608 
#>  4 Triclosan 67606152 
#>  5 Triclosan 71752714 
#>  6 Triclosan 24848164 
#>  7 Triclosan 67346638 
#>  8 Triclosan 45040609 
#>  9 Triclosan 131741682
#> 10 Triclosan 67606151 
#> # … with 30 more rows

# from SMILES
get_cid("CCCC", from = "smiles")

#> # A tibble: 1 x 2
#>   query cid  
#>   <chr> <chr>
#> 1 CCCC  7843 

# from InChI
get_cid("InChI=1S/CH5N/c1-2/h2H2,1H3", from = "inchi")

#> # A tibble: 1 x 2
#>   query                       cid  
#>   <chr>                       <chr>
#> 1 InChI=1S/CH5N/c1-2/h2H2,1H3 6329 

# from InChIKey
get_cid("BPGDAMSIGCZZLK-UHFFFAOYSA-N", from = "inchikey")

#> # A tibble: 1 x 2
#>   query                       cid  
#>   <chr>                       <chr>
#> 1 BPGDAMSIGCZZLK-UHFFFAOYSA-N 12345

# from formula
get_cid("C26H52NO6P", from = "formula")

#> # A tibble: 33 x 2
#>    query      cid      
#>    <chr>      <chr>    
#>  1 C26H52NO6P 10864091 
#>  2 C26H52NO6P 24779513 
#>  3 C26H52NO6P 71584579 
#>  4 C26H52NO6P 71582867 
#>  5 C26H52NO6P 6384103  
#>  6 C26H52NO6P 4465356  
#>  7 C26H52NO6P 147059762
#>  8 C26H52NO6P 131716652
#>  9 C26H52NO6P 91019382 
#> 10 C26H52NO6P 88696684 
#> # … with 23 more rows

# from CAS RN
get_cid("56-40-6", from = "xref/rn")

#> # A tibble: 6 x 2
#>   query   cid     
#>   <chr>   <chr>   
#> 1 56-40-6 750     
#> 2 56-40-6 22316   
#> 3 56-40-6 40772   
#> 4 56-40-6 5257127 
#> 5 56-40-6 21226085
#> 6 56-40-6 23673715

# similarity
get_cid(5564, from = "similarity/cid")

#> # A tibble: 197 x 2
#>    query cid     
#>    <dbl> <chr>   
#>  1  5564 5564    
#>  2  5564 627458  
#>  3  5564 5271320 
#>  4  5564 107942  
#>  5  5564 18807   
#>  6  5564 21272512
#>  7  5564 24848164
#>  8  5564 9822749 
#>  9  5564 165111  
#> 10  5564 67606152
#> # … with 187 more rows

get_cid("CCO", from = "similarity/smiles")

#> # A tibble: 192 x 2
#>    query cid     
#>    <chr> <chr>   
#>  1 CCO   702     
#>  2 CCO   12201693
#>  3 CCO   12201695
#>  4 CCO   12201684
#>  5 CCO   519451  
#>  6 CCO   137238  
#>  7 CCO   123157  
#>  8 CCO   123093  
#>  9 CCO   102138  
#> 10 CCO   58233034
#> # … with 182 more rows

# from SID
get_cid("126534046", from = "sid", domain = "substance")

#> # A tibble: 1 x 2
#>   query     cid     
#>   <chr>     <chr>   
#> 1 126534046 24971898

# sourceid
get_cid("VCC957895", from = "sourceid/23706", domain = "substance")

#> # A tibble: 1 x 2
#>   query     cid     
#>   <chr>     <chr>   
#> 1 VCC957895 19689584

# sourceall
get_cid("Optopharma Ltd", from = "sourceall", domain = "substance")

#> # A tibble: 3 x 2
#>   query          cid      
#>   <chr>          <chr>    
#> 1 Optopharma Ltd 155511990
#> 2 Optopharma Ltd 102361739
#> 3 Optopharma Ltd 129626534

# from AID (CIDs of substances tested in the assay)
get_cid(170004, from = "aid", domain = "assay")

#> # A tibble: 1 x 2
#>    query cid  
#>    <dbl> <chr>
#> 1 170004 68352

# from GeneID (CIDs of substances tested on the gene)
get_cid(25086, from = "target/geneid", domain = "assay")

#> # A tibble: 25 x 2
#>    query cid     
#>    <dbl> <chr>   
#>  1 25086 11659928
#>  2 25086 11624320
#>  3 25086 11580958
#>  4 25086 11595347
#>  5 25086 11610203
#>  6 25086 11681690
#>  7 25086 44565544
#>  8 25086 25143210
#>  9 25086 44565496
#> 10 25086 44565517
#> # … with 15 more rows

# multiple inputs
get_cid(c("Triclosan", "Aspirin"))

#> # A tibble: 2 x 2
#>   query     cid  
#>   <chr>     <chr>
#> 1 Triclosan 5564 
#> 2 Aspirin   2244 

# }