Retrieve data from PubChem content pages

When you search for an entity at https://pubchem.ncbi.nlm.nih.gov/, e.g. a compound or a substance, and select the record you are interested in, you will be forwarded to a PubChem content page. When you look at a PubChem content page, you can see that chemical information is organised into sections, subsections, etc. The chemical data live at the lowest levels of these sections. Use this function to retrieve the lowest level information from PubChem content pages.

pc_sect(
  id,
  section,
  domain = c("compound", "substance", "assay", "gene", "protein", "patent"),
  verbose = getOption("verbose")
)

Arguments

id
numeric or character; a vector of PubChem identifiers to search for.

section
character; the section of the content page to be imported.

domain
character; the query domain. Can be one of "compound", "substance", "assay", "gene", "protein" or "patent".

verbose
logical; should a verbose output be printed on the console?

Value

Returns a tibble of query results. In the returned tibble, SourceName is the name of the depositor, and SourceID is the ID of the search term within the depositor's database. You can browse https://pubchem.ncbi.nlm.nih.gov/sources/ for more information about the depositors.

Details

section is not case sensitive but it is sensitive to typing errors and it requires the full name of the section as it is printed on the content page. The PubChem Table of Contents Tree can also be found at https://pubchem.ncbi.nlm.nih.gov/classification/#hid=72.

Note

Please respect the Terms and Conditions of the National Library of Medicine, https://www.nlm.nih.gov/databases/download.html the data usage policies of National Center for Biotechnology Information, https://www.ncbi.nlm.nih.gov/home/about/policies/, https://pubchemdocs.ncbi.nlm.nih.gov/programmatic-access, and the data usage policies of the individual data sources https://pubchem.ncbi.nlm.nih.gov/sources/.

References

Kim, S., Thiessen, P.A., Cheng, T. et al. PUG-View: programmatic access to chemical annotations integrated in PubChem. J Cheminform 11, 56 (2019). doi: 10.1186/s13321-019-0375-2 .

See also

Examples

# might fail if API is not available
# \donttest{
pc_sect(176, "pka")

#> # A tibble: 1 x 5
#>   CID   Name        Result          SourceName SourceID
#>   <chr> <chr>       <chr>           <chr>      <chr>   
#> 1 176   Acetic acid 4.76 (at 25 °C) DrugBank   DB03166 

pc_sect(c(176, 311), "density")

#> # A tibble: 12 x 5
#>    CID   Name    Result                      SourceName                 SourceID
#>    <chr> <chr>   <chr>                       <chr>                      <chr>   
#>  1 176   Acetic… 1.051 at 68 °F (USCG, 1999) CAMEO Chemicals            2272    
#>  2 176   Acetic… 1.0446 g/cu cm at 25 °C     Hazardous Substances Data… 40      
#>  3 176   Acetic… Density: 1.266 at 16.60 °C… Hazardous Substances Data… 40      
#>  4 176   Acetic… Relative density (water = … ILO International Chemica… 0363    
#>  5 176   Acetic… 1.049                       Joint FAO/WHO Expert Comm… f_3     
#>  6 176   Acetic… 1.05                        Occupational Safety and H… 469     
#>  7 176   Acetic… 1.05                        The National Institute fo… npgd0002
#>  8 311   Citric… 1.54 at 68 °F (USCG, 1999)  CAMEO Chemicals            10899   
#>  9 311   Citric… 1.665 g/cu cm at 20 °C      Hazardous Substances Data… 911     
#> 10 311   Citric… BULK DENSITY: 56.2 LB/CU F… Hazardous Substances Data… 911     
#> 11 311   Citric… Density: 1.542 g/cu cm /Ci… Hazardous Substances Data… 911     
#> 12 311   Citric… White, odorless crystals, … Hazardous Substances Data… 911     

pc_sect(2231, "depositor-supplied synonyms", "substance")

#> # A tibble: 3 x 5
#>   SID   Name        Result                SourceName SourceID
#>   <chr> <chr>       <chr>                 <chr>      <chr>   
#> 1 2231  cholesterol cholesterol           PubChem    PubChem 
#> 2 2231  cholesterol 57-88-5               PubChem    PubChem 
#> 3 2231  cholesterol 5-cholestene-3beta-ol PubChem    PubChem 

pc_sect(780286, "modify date", "assay")

#> # A tibble: 3 x 5
#>   AID    Name                                       Result   SourceName SourceID
#>   <chr>  <chr>                                      <chr>    <chr>      <chr>   
#> 1 780286 Reduction in [3H]cholesterol content in e… 2014-05… PubChem    PubChem 
#> 2 780286 Reduction in [3H]cholesterol content in e… 2018-09… PubChem    PubChem 
#> 3 780286 Reduction in [3H]cholesterol content in e… 2020-06… PubChem    PubChem 

pc_sect(9023, "Ensembl ID", "gene")

#> # A tibble: 1 x 5
#>   GeneID Name                                  Result        SourceName SourceID
#>   <chr>  <chr>                                 <chr>         <chr>      <chr>   
#> 1 9023   CH25H - cholesterol 25-hydroxylase (… ENSG00000138… NCBI Gene  9023    

pc_sect("1ZHY_A", "Sequence", "protein")

#> # A tibble: 2 x 5
#>   pdbID  Name          Result                                SourceName SourceID
#>   <chr>  <chr>         <chr>                                 <chr>      <chr>   
#> 1 1ZHY_A Chain A, KES… >pdb|1ZHY|A Chain A, KES1 protein (R… NCBI Prot… 1ZHY_A  
#> 2 1ZHY_A Chain A, KES… GAMDPSQYASSSSWTSFLKSIASFNGDLSSLSAPPF… NCBI Prot… 1ZHY_A  

# }